added pdf bills after order

+++ src/HotelManagementApp.Application/Responses/RoomResponses/RoomResponse.cs
@@ -6,7 +6,7 @@ public class RoomResponse
     public required string RoomName { get; set; }
     public required string RoomType { get; set; }
     public required decimal Price { get; set; }
-    public int DiscountPercent { get; set; }
+    public decimal DiscountPercent { get; set; }
     public required decimal FinalPrice { get; set; }
     public required int HotelId { get; set; }
     public ICollection<string> RoomImages { get; set; } = new List<string>();
